Refactor

+++ src/main/resources/view/modules/TopDownGameModule.js
@@ -5,12 +5,14 @@ export class TopDownGameModule {
 
     constructor() {
         this.mapContainer = null;
+        this.characterContainer = null;
         this.conversationHelper = new Conversation(PIXI);
         this.noFocusHelper = new NoFocus(PIXI);
+        this.activeAreas = new Set();
+        this.keys = {};
     }
 
-    keys = {};
-
+    // region map
     showMap(tileset, mapData, tileSize, antialias, mapScale, mapX, mapY) {
         // load tileset asset and set/unset antialiasing
         const base = PIXI.BaseTexture.from(tileset);
@@ -28,7 +30,6 @@ export class TopDownGameModule {
             return new PIXI.Texture(tileTexture.baseTexture, frame);
         };
 
-        this.mapContainer.sortableChildren = true;
         for (let layer = 0; layer < mapData.length; layer++) {
             for (let row = 0; row < mapData[0].length; row++) {
                 for (let col = 0; col < mapData[0][0].length; col++) {
@@ -59,30 +60,20 @@ export class TopDownGameModule {
 
         this.mapContainer.x = mapX;
         this.mapContainer.y = mapY;
+        this.characterContainer.x = mapX;
+        this.characterContainer.y = mapY;
     }
 
-    showCharacter(character, tileSize, mapScale) {
-        const sprite = character.sprite; // assets/$sprite is relative path to PNG sprite
-        const charX = character.x;
-        const charY = character.y;
-
-        this.character = new PIXI.Sprite.from(sprite);
+    showCharacter(tileSize, mapScale) {
+        this.character = new PIXI.Sprite.from(this.frameData.character.sprite);
         this.character.texture.baseTexture.scaleMode = PIXI.SCALE_MODES.NEAREST;
         this.character.roundPixels = true;
-        this.character.position.set(charX * tileSize * mapScale, charY * tileSize * mapScale);
+        this.character.position.set(
+            this.frameData.character.x * tileSize * mapScale,
+            this.frameData.character.y * tileSize * mapScale
+        );
         this.character.scale.set(mapScale);
-
-        this.mapContainer.addChild(this.character);
-    }
-
-    setupGameZones(obstacles, areas, conversations) {
-        this.obstacles = obstacles;
-        this.areas = areas;
-        this.conversations = conversations;
-        this.textContainer = new PIXI.Container();
-        this.container.addChild(this.textContainer);
-
-        this.conversationHelper.setUp(conversations, this.textContainer);
+        this.characterContainer.addChild(this.character);
     }
 
     showLevel() {
@@ -97,47 +88,28 @@ export class TopDownGameModule {
         const mapX = this.frameData.x;
         const mapY = this.frameData.y;
 
-        console.log(this.frameData);
-
         this.totalTileSize = mapScale * tileSize;
 
-        // unused yet
-        const obstacles = this.frameData.tileMapObstacles;
-        const areas = this.frameData.areas;
-        const conversations = this.frameData.conversations;
-        const character = this.frameData.character;
-
         this.showMap(tileset, mapData, tileSize, antialias, mapScale, mapX, mapY);
-        this.showCharacter(character, tileSize, mapScale);
-        this.setupGameZones(obstacles, areas, conversations);
+        this.showCharacter(tileSize, mapScale);
+
+        this.conversationHelper.init(this.frameData.conversations, this.conversationContainer);
+
+        this.mapContainer.cacheAsBitmap = true;
+    }
+    // endregion
+
+    // region collisions
+    blocked() {
+        const col = this.character.x / this.totalTileSize;
+        const row = this.character.y / this.totalTileSize;
+        return this.frameData.tileMapObstacles[Math.floor(row)][Math.floor(col)] === true ||
+            this.frameData.tileMapObstacles[Math.ceil(row)][Math.floor(col)] === true ||
+            this.frameData.tileMapObstacles[Math.floor(row)][Math.ceil(col)] === true ||
+            this.frameData.tileMapObstacles[Math.ceil(row)][Math.ceil(col)] === true
     }
 
-    handleFrameData(frameInfo, frameData) {
-        if (frameData) { this.frameData = frameData; }
-        this.showLevel();
-        this.noFocusHelper.initNoFocusWindow(this.noFocusContainer, this.frameData);
-    }
-
-    handleGlobalData(players, globalData) { }
-
-    updateScene(previousData, currentData, progress, speed) { }
-
-    reinitScene(container, canvasData) {
-        this.mapContainer = new PIXI.Container();
-        this.mapContainer.sortableChildren = true;
-        this.noFocusContainer = new PIXI.Container();
-        this.container = container;
-        container.addChild(this.mapContainer);
-        container.addChild(this.noFocusContainer);
-
-        this.showLevel();
-        this.noFocusHelper.initNoFocusWindow(this.noFocusContainer, this.frameData);
-
-        window.addEventListener('keydown', (e) => { this.keys[e.code] = true; this.noFocusHelper.forceFocus(true); });
-        window.addEventListener('keyup', (e) => { this.keys[e.code] = false; this.noFocusHelper.forceFocus(true); });
-    }
-
-    handleMovement(delta) {
+    handleMovement() {
         const moveSpeed = 5;
 
         const originalX = this.character.x;
@@ -145,41 +117,17 @@ export class TopDownGameModule {
 
         if (this.keys.KeyW) this.character.y -= moveSpeed;
         if (this.keys.KeyS) this.character.y += moveSpeed;
-
-        const col1 = this.character.x / this.totalTileSize;
-        const row1 = this.character.y / this.totalTileSize;
-
-        if (
-            this.obstacles[Math.floor(row1)][Math.floor(col1)] === true ||
-            this.obstacles[Math.ceil(row1)][Math.floor(col1)] === true ||
-            this.obstacles[Math.floor(row1)][Math.ceil(col1)] === true ||
-            this.obstacles[Math.ceil(row1)][Math.ceil(col1)] === true
-        ) {
-            this.character.y = originalY;
-        }
+        if (this.blocked()) { this.character.y = originalY; }
 
         if (this.keys.KeyA) this.character.x -= moveSpeed;
         if (this.keys.KeyD) this.character.x += moveSpeed;
-
-        const col2 = this.character.x / this.totalTileSize;
-        const row2 = this.character.y / this.totalTileSize;
-
-        if (
-            this.obstacles[Math.floor(row2)][Math.floor(col2)] === true ||
-            this.obstacles[Math.ceil(row2)][Math.floor(col2)] === true ||
-            this.obstacles[Math.floor(row2)][Math.ceil(col2)] === true ||
-            this.obstacles[Math.ceil(row2)][Math.ceil(col2)] === true
-        ){
-            this.character.x = originalX;
-        }
-
-        if (this.character.x !== originalX || this.character.y !== originalY) {
-            this.handleConversationZones();
-        }
+        if (this.blocked()) { this.character.x = originalX; }
     }
+    // endregion
 
-    handleConversationZones() {
-        if (!this.character || !this.areas) return;
+    // region areas
+    resolveAreaVisited() {
+        if (!this.frameData) return;
 
         const charX = this.character.x;
         const charY = this.character.y;
@@ -191,7 +139,7 @@ export class TopDownGameModule {
         const charTop = charY;
         const charBottom = charY + charH;
 
-        for (const [id, area] of Object.entries(this.areas)) {
+        for (const [id, area] of Object.entries(this.frameData.areas)) {
             if (area.type !== 'RECTANGLE') continue;
 
             const areaLeft = area.ax * this.totalTileSize;
@@ -206,29 +154,73 @@ export class TopDownGameModule {
                 charTop <= areaBottom;
 
             if (intersects) {
-                this.onAreaEntered(id);
+                if (!this.activeAreas.has(id)) {
+                    this.activeAreas.add(id);
+                    this.onAreaEntered(id);
+                }
             } else {
-                this.onAreaLeft(id);
+                if (this.activeAreas.has(id)) {
+                    this.activeAreas.delete(id);
+                    this.onAreaLeft(id);
+                }
             }
         }
     }
 
-    onAreaEntered(id) {
-        this.conversationHelper.enableConversation(id);
+    onAreaEntered(id) { this.conversationHelper.initConversation(id); }
+    onAreaLeft(id) { this.conversationHelper.leaveConversation(id); }
+    // endregion
+
+    /** CG API - handle data for current frame */
+    handleFrameData(frameInfo, frameData) {
+        if (frameData) { this.frameData = frameData; }
+        this.showLevel();
+        this.noFocusHelper.init(this.noFocusContainer, this.frameData);
     }
 
-    onAreaLeft(id) { this.conversationHelper.leaveConversation(); }
+    /** CG API */
+    handleGlobalData(players, globalData) { }
 
+    /** CG API */
+    updateScene(previousData, currentData, progress, speed) { }
+
+    /** CG API - called when scene needs reinit - including resising. */
+    reinitScene(container, canvasData) {
+        // TODO: persist character position
+        this.mapContainer = new PIXI.Container();
+        this.characterContainer = new PIXI.Container();
+        this.mapContainer.sortableChildren = true;
+        this.noFocusContainer = new PIXI.Container();
+        this.conversationContainer = new PIXI.Container();
+
+        container.addChild(this.mapContainer);
+        container.addChild(this.characterContainer);
+        container.addChild(this.noFocusContainer);
+        container.addChild(this.conversationContainer);
+
+        this.showLevel();
+        this.resolveAreaVisited();
+
+        this.noFocusHelper.init(this.noFocusContainer, this.frameData);
+
+        window.addEventListener('keydown', (e) => { this.keys[e.code] = true; this.noFocusHelper.forceFocus(true); });
+        window.addEventListener('keyup', (e) => { this.keys[e.code] = false; this.noFocusHelper.forceFocus(true); });
+    }
+
+    /** CG API - called every few ms - ideal for key press checks for fluent movement */
     animateScene(delta) {
-        if (!this.character) return;
+        if (!this.frameData) return;
         if (!this.prevKeys) { this.prevKeys = {}; }
 
+        // key clicked -> try to process conversation
         if (JSON.stringify(this.prevKeys) !== JSON.stringify(this.keys)) {
             this.conversationHelper.processKeys(this.keys);
         }
 
+        // key held -> try to process movement
         if (Object.values(this.keys).some(v => v === true)) {
-            this.handleMovement(delta);
+            this.handleMovement();
+            this.resolveAreaVisited();
         }
 
         this.prevKeys = { ...this.keys };
@@ -236,6 +228,23 @@ export class TopDownGameModule {
 
 }
 
+/**
+ * This class is meant as overlay over the game while game is not in focus. Currently, it requires these
+ * attributed from FrameData (all shown while game is not in focus - hidden otherwise)
+ *
+ * - noFocusOverlayAsset - relative path of asset displayed in center of the screen
+ * - noFocusOverlayBackgroundColor - int color shown as background
+ * - noFocusOverlayBackgroundAlpha - transparency of background color
+ *
+ * Initialize this overlay using initNoFocusWindow() when you have PIXI container and FrameData available
+ *
+ * This overlay is triggered using these triggers
+ * - window.onfocus
+ * - window.onblur
+ *
+ * Or you can show it manually using
+ * - forceFocus(true/false)
+ */
 class NoFocus {
 
     constructor(PIXI) {
@@ -248,7 +257,7 @@ class NoFocus {
         this.invalidateFocus();
     }
 
-    initNoFocusWindow(container, frameData) {
+    init(container, frameData) {
         if (!container || !frameData) return;
 
         container.removeChildren().forEach(c => c.destroy({ children: true }));
@@ -278,6 +287,14 @@ class NoFocus {
     }
 }
 
+/**
+ * This class is meant as overlay over the game to show conversations.
+ *
+ * - init() - to initialize conversations with PIXI container
+ * - initConversation(id) - to initialize conversation (either immediately or showing "press T to talk")
+ * - leaveConversation(id) - to leave conversation with given ID - should be paired with initConversation
+ * - processKeys(keys) - to process keys related to conversation
+ */
 class Conversation {
 
     constructor(PIXI) {
@@ -301,50 +318,59 @@ class Conversation {
         });
     }
 
-    setUp(conversations, pixiContainer) {
+    init(conversations, pixiContainer) {
+        if (!conversations || !pixiContainer) return;
         this.conversations = conversations;
         this.talkTextContainer = pixiContainer;
     }
 
-    enableConversation(id) {
-        this.conversationId = id;
-        this.talkTextContainer.removeChildren();
-        const container = new this.PIXI.Container();
+    initConversation(id) {
+        if (this.conversationId === id) return;
 
-        let offsetX = 0;
-
-        const t1 = new this.PIXI.Text("Press ", this.baseTextStyle);
-        t1.x = offsetX;
-        container.addChild(t1);
-        offsetX += t1.width;
-
-        const t2 = new this.PIXI.Text("[T]", this.boldTextStyle);
-        t2.x = offsetX;
-        container.addChild(t2);
-        offsetX += t2.width;
-
-        const t3 = new this.PIXI.Text(" to talk.", this.baseTextStyle);
-        t3.x = offsetX;
-        container.addChild(t3);
-
-        this.talkTextContainer.addChild(container);
-        container.x = 50;
-        container.y = 100;
-    }
-
-    startConversation() {
-        if (!this.currentConversation) {
-            this.currentConversation = this.conversations[this.conversationId];
-            this.currentLineIndex = 0;
-            this.renderConversationLine();
+        if (this.conversations[id].immediate) {
+            this.startConversation(id);
+        } else {
+            this.showPressToTalk(id);
         }
     }
 
-    leaveConversation() {
+    showPressToTalk(id) {
+        if (this.isPressToTalkVisible) return;  // already visible
+        if (this.conversationId === id) return; // conversation already running
+        this.conversationId = id;
+        this.conversationStarted = false;
+
+        this.talkTextContainer.removeChildren();
+        const container = new this.PIXI.Container();
+
+        const t1 = new this.PIXI.Text("Press [T] to talk.", this.boldTextStyle);
+        container.addChild(t1);
+
+        this.talkTextContainer.addChild(container);
+        container.x = 50;
+        container.y = 50;
+    }
+
+    startConversation(id) {
+        if (this.conversationId === id && this.conversationStarted) return;
+
+        this.conversationId = id;
+        this.conversationStarted = true;
+        this.currentConversation = this.conversations[this.conversationId];
+        this.currentLineIndex = 0;
+        this.isPressToTalkVisible = false;
+        this.renderConversationLine();
+    }
+
+    leaveConversation(id) {
+        if (id !== this.conversationId) { return; }
+
         this.talkTextContainer.removeChildren().forEach(c => c.destroy({children: true}));
         this.conversationId = null;
+        this.conversationStarted = false;
         this.currentConversation = null;
-        this.currentLineIndex = -1;
+        this.currentLineIndex = null;
+        this.isPressToTalkVisible = false;
     }
 
     renderConversationLine() {
@@ -379,9 +405,10 @@ class Conversation {
 
     nextConversationLine() {
         if (!this.currentConversation) return;
+
         this.currentLineIndex++;
-        if (this.currentLineIndex === -1 || this.currentLineIndex >= this.currentConversation.content.length) {
-            this.leaveConversation()
+        if (this.currentLineIndex >= this.currentConversation.content.length) {
+            this.leaveConversation(this.conversationId);
         } else {
             this.renderConversationLine();
         }
@@ -392,7 +419,7 @@ class Conversation {
             if (this.currentConversation) {
                 this.nextConversationLine();
             } else {
-                this.startConversation();
+                this.startConversation(this.conversationId);
             }
         }
     }
